requirements.txt
Django==4.2.7
django-tenant-schemas==1.10.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-import-export==3.3.1
django-filter==23.3
django-phonenumber-field==7.1.0
phonenumbers==8.13.23
Pillow==11.2.1
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
django-mptt==0.15.0
django-taggit==4.0.0
django-summernote==0.8.20.0
django-bootstrap-datepicker-plus==5.0.4
django-widget-tweaks==1.5.0
reportlab==4.0.5
xlsxwriter==3.1.9
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.0
django-storages==1.14.2
boto3==1.29.7
gunicorn==23.0.0
whitenoise==6.9.0
sentry-sdk==1.38.0
django-environ==0.11.2
django-cacheops==7.0.2
django-silk==5.0.4
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-countries==7.5.1
django-timezone-field==6.0.1
python-dateutil==2.8.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0